Moved fetch to app.jsx.
