+++ src/pages/MovieList.jsx
@@ -1,7 +1,7 @@
 //import { Link } from "react-router-dom"
 import { Movie } from "../components/Movie"
 //import ships from "../star_wars_ships.json"
-import { FetchMovieDetails } from "../components/FetchMovieDetails"
+//import { FetchMovieDetails } from "../components/FetchMovieDetails"
 import styled from "styled-components"
 
 const MovieListWrapper = styled.div`
@@ -10,16 +10,12 @@ const MovieListWrapper = styled.div`
   justify-content: center;
   `
 
-export const MovieList = ({movie}) => {
+export const MovieList = ( { movie }) => {
 
 
-  
-
-return (
+  return (
     <>
-    <FetchMovieDetails />
-      {/* <h1>The Popular Movies</h1>  */}
-      <MovieListWrapper>
+    <MovieListWrapper>
         {movie.map((movie) => (
           <Movie
             key={movie.id}
@@ -29,18 +25,6 @@ return (
           />
         ))}
       </MovieListWrapper>
-
     </>
   )
-
-
-//   return (
-//     <>
-// {ships.map((ship) => (
-// <a href="">
-// <Movie name={ship.name} image={ship.image_url} /> 
-// </a>
-// ) )}
-//     </>
-//   )
 }
